Refactor tests further by functionality.
